fixed static analyzer complaints
