+++ libs/ml/include/ml/serializers/ml_types.hpp
@@ -29,33 +29,6 @@ namespace ml {
 template <typename T>
 struct StateDict;
 
-namespace model {
-template <typename TensorType>
-class Model;
-template <typename TensorType>
-class Sequential;
-template <typename DataType>
-struct ModelConfig;
-}  // namespace model
-
-namespace dataloaders {
-template <typename TensorType>
-class TensorDataLoader;
-}  // namespace dataloaders
-
-namespace optimisers {
-template <typename TensorType>
-class SGDOptimiser;
-template <typename TensorType>
-class AdamOptimiser;
-}  // namespace optimisers
-
-namespace utilities {
-template <typename T>
-void BuildGraph(fetch::ml::GraphSaveableParams<T> const &sp,
-                std::shared_ptr<fetch::ml::Graph<T>>     ret);
-}  // namespace utilities
-
 }  // namespace ml
 
 namespace serializers {
@@ -3514,90 +3487,6 @@ struct MapSerializer<ml::LayerSkipGramSaveableParams<TensorType>, D>
     map.ExpectKeyGetValue(IN_SIZE, sp.in_size);
     map.ExpectKeyGetValue(OUT_SIZE, sp.out_size);
     map.ExpectKeyGetValue(EMBED_IN, sp.embed_in);
-  }
-};
-
-/**
- * serializer for Sequential model
- * @tparam TensorType
- */
-template <typename TensorType, typename D>
-struct MapSerializer<ml::model::Sequential<TensorType>, D>
-{
-  using Type                          = ml::model::Sequential<TensorType>;
-  using DriverType                    = D;
-  static uint8_t const BASE_MODEL     = 1;
-  static uint8_t const LAYER_COUNT    = 2;
-  static uint8_t const PREV_LAYER_STR = 3;
-
-  template <typename Constructor>
-  static void Serialize(Constructor &map_constructor, Type const &sp)
-  {
-    auto map = map_constructor(3);
-
-    // serialize the optimiser parent class
-    auto base_pointer = static_cast<ml::model::Model<TensorType> const *>(&sp);
-    map.Append(BASE_MODEL, *base_pointer);
-    map.Append(LAYER_COUNT, sp.layer_count_);
-    map.Append(PREV_LAYER_STR, sp.prev_layer_);
-  }
-
-  template <typename MapDeserializer>
-  static void Deserialize(MapDeserializer &map, Type &sp)
-  {
-    auto base_pointer = static_cast<ml::model::Model<TensorType> *>(&sp);
-    map.ExpectKeyGetValue(BASE_MODEL, *base_pointer);
-    map.ExpectKeyGetValue(LAYER_COUNT, sp.layer_count_);
-    map.ExpectKeyGetValue(PREV_LAYER_STR, sp.prev_layer_);
-  }
-};
-
-/**
- * serializer for ModelConfig
- * @tparam TensorType
- */
-template <typename DataType, typename D>
-struct MapSerializer<fetch::ml::model::ModelConfig<DataType>, D>
-{
-  using Type       = fetch::ml::model::ModelConfig<DataType>;
-  using DriverType = D;
-
-  // public member variables
-  static uint8_t const EARLY_STOPPING      = 1;
-  static uint8_t const TEST                = 2;
-  static uint8_t const PATIENCE            = 3;
-  static uint8_t const MIN_DELTA           = 4;
-  static uint8_t const LEARNING_RATE_PARAM = 6;
-  static uint8_t const BATCH_SIZE          = 7;
-  static uint8_t const SUSBET_SIZE         = 8;
-  static uint8_t const PRINT_STATS         = 9;
-
-  template <typename Constructor>
-  static void Serialize(Constructor &map_constructor, Type const &sp)
-  {
-    auto map = map_constructor(9);
-
-    map.Append(EARLY_STOPPING, sp.early_stopping);
-    map.Append(TEST, sp.test);
-    map.Append(PATIENCE, sp.patience);
-    map.Append(MIN_DELTA, sp.min_delta);
-    map.Append(LEARNING_RATE_PARAM, sp.learning_rate_param);
-    map.Append(BATCH_SIZE, sp.batch_size);
-    map.Append(SUSBET_SIZE, sp.subset_size);
-    map.Append(PRINT_STATS, sp.print_stats);
-  }
-
-  template <typename MapDeserializer>
-  static void Deserialize(MapDeserializer &map, Type &sp)
-  {
-    map.ExpectKeyGetValue(EARLY_STOPPING, sp.early_stopping);
-    map.ExpectKeyGetValue(TEST, sp.test);
-    map.ExpectKeyGetValue(PATIENCE, sp.patience);
-    map.ExpectKeyGetValue(MIN_DELTA, sp.min_delta);
-    map.ExpectKeyGetValue(LEARNING_RATE_PARAM, sp.learning_rate_param);
-    map.ExpectKeyGetValue(BATCH_SIZE, sp.batch_size);
-    map.ExpectKeyGetValue(SUSBET_SIZE, sp.subset_size);
-    map.ExpectKeyGetValue(PRINT_STATS, sp.print_stats);
   }
 };
